Revert "Set up Jest for content editor testing"
This reverts commit 7749d2c5aabdabeffdae298e7ae05cc645184b72.
