--- babel.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-    presets: [
-      ['@babel/preset-env', {targets: {node: 'current'}}],
-      '@babel/preset-typescript',
-      "@babel/preset-react",
-    ],
-    plugins: [
-      "@babel/plugin-transform-react-jsx"
-    ],
-  };
