+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@babel/plugin-transform-react-jsx": "^7.25.9",
     "@codemirror/lang-json": "^0.20.0",
     "@codemirror/lang-markdown": "^0.20.1",
     "@testing-library/jest-dom": "^5.16.1",
@@ -85,7 +84,6 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@babel/preset-typescript": "^7.26.0",
     "@craco/craco": "^7.1.0",
     "@craco/types": "^7.1.0",
     "@types/he": "^1.1.2",
