+++ src/services/llmMarkingFormula.ts
@@ -1,6 +1,7 @@
 import { Parser, Grammar } from 'nearley';
 import grammar from '../grammars/llmMarkingFormula.nearley';
 
+
 export function parseMarkingFormula(markingFormula?: string) {
     if (!markingFormula) {
         return "";
